python-inference/requirements.txt
redis
Pillow
Flask
opencv-python-headless
numpy